[Refactor]: Migrate claim function to better context
